+++ client/src/lib/parse/timeline/TimelineParser.ts
@@ -73,6 +73,7 @@ export default class TimelineParser {
   static addSearchTagToFacetValues(
     items: Array<IOrderedItems>,
     searchTag: string,
+    halLink: string,
   ): Array<ITransformedData> {
     const transformedFacets: Array<ITransformedData> = []
     for (const item of items) {
@@ -83,6 +84,7 @@ export default class TimelineParser {
           totalItems: totalItems as number,
           searchTag,
           id: this.convertIdToSearchQueryParams(id),
+          halLink,
         })
       }
     }
@@ -187,19 +189,19 @@ export default class TimelineParser {
    * @param {ITimelinesTransformed} timelineData; the transformed timeline data
    * @returns {Array<string>}
    */
-  static getFacetsUsedForLegend(
+  static getHalLinksUsedForLegend(
     timelineData: ITimelinesTransformed,
   ): Array<string> {
-    const facetKeys: Array<string> = []
+    const halLinkKeys: Array<string> = []
     for (const yearKey of Object.keys(timelineData)) {
       const yearData = timelineData[yearKey]
-      for (const facetKey of Object.keys(yearData)) {
-        if (facetKey !== 'total' && !facetKeys.includes(facetKey)) {
-          facetKeys.push(facetKey)
+      for (const halLinkKey of Object.keys(yearData)) {
+        if (halLinkKey !== 'total' && !halLinkKeys.includes(halLinkKey)) {
+          halLinkKeys.push(halLinkKey)
         }
       }
     }
-    return facetKeys
+    return halLinkKeys
   }
 
   /**
@@ -219,44 +221,47 @@ export default class TimelineParser {
     for (const result of this.timeline) {
       // the key is the api endpoint to retreive the facet values
       for (const key of Object.keys(result)) {
-        const { orderedItems } = result[key]
-        const searchTag = TimelineParser.getSearchTagFromFacetedSearch(key)
-
+        const { orderedItems, id } = result[key]
+        const searchTag = TimelineParser.getSearchTagFromFacetedSearch(id)
         if (orderedItems !== null && orderedItems.length > 0) {
           transformedData = [
             ...transformedData,
             ...TimelineParser.addSearchTagToFacetValues(
               orderedItems,
               searchTag,
+              key,
             ),
           ]
         }
       }
     }
-
+    // filter out any null values for years
     transformedData.filter((value) => value !== null)
 
     const dateCounts: { [key: string]: any } = {}
     for (const tData of transformedData) {
-      const { value, totalItems, searchTag, id } = tData as ITransformedData
+      const { value, totalItems, searchTag, id, halLink } =
+        tData as ITransformedData
       const date = String(value)
       const individualDate = dateCounts[date]
       if (dateCounts.hasOwnProperty(date)) {
         individualDate.total += totalItems
-        if (!individualDate.hasOwnProperty(searchTag)) {
-          individualDate[searchTag] = {
+        if (!individualDate.hasOwnProperty(halLink)) {
+          individualDate[halLink] = {
             totalItems,
             searchParams: id,
+            searchTag,
           }
         } else {
-          dateCounts[date][searchTag].totalItems += totalItems
+          dateCounts[date][halLink].totalItems += totalItems
         }
       } else {
         dateCounts[date] = {
           total: totalItems,
-          [searchTag]: {
+          [halLink]: {
             totalItems,
             searchParams: id,
+            searchTag,
           },
         }
       }